+++ src/Sql/Sql/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for restore of dropped databases on Managed Instances.
+
+## Version 1.15.0
+* Add support for setting Active Directory Administrator on Managed Instance
 
 ## Version 1.14.2
 * Update example in reference documentation for `Get-AzSqlElasticPool`